Issue 337: TAMU Header should support the suppression of the top navigation links, both complete suppression and discriminate.
Utilize ngIf to conditionally suppress the top nav.
This avoids making any decisions on whether or not ngIf should be used for performance and memory reasons.

Implement a simple array to determine whether or not a given named element is to be suppressed.
The word 'all' is reserved for suppressing all navigation links.

The naming system is not explicitly provided so I opted to lowercase and trim the title text.
This is done because the user is most likely to know the title text.
This is only a noticeable problem for `information for`.
Other options are to always strip whitespace, or replace all inner whitespace with a '-', such as `information-for`.