Fix problem introduced by following linter without fully converting loop.
The linter complained about the loop `for key in ..`.
The loop was changed to a `for key of ..`.
The problem happens because the `values[key]` should have also been changed to `key` but was not.
This resulted in a regression due to incompletely converting the loop type.
This is now fixed and `name` is used rather than `key`.