Issue 395: Verify that TAMU Library Components is showing icons in the mobile menu.
The issue is verified that the issue appears only in 1.x and not 2.x.
Investigating why this is the case revealed that the icons aren't even in the repository.
The 2.x prevents blacklisting the bootstrap icons assets in the `.gitignore` file but the 1.x does not.

This issue is tagged as a spike, but fixing the issue immediately makes more sense given that it is just adding static files.