Add `npm-local` command to pull in `wvr/elements` without building.
The `build:npm-local` work for starting the dev environment but not for the dist.
This is because it doesn't build the bundle
The `build` command will build the bundle that is needed by `run:dist`.
Rather than modify `build`, just provide another command that can be run as-is without being locked into any specific build process.