Change tl nav list and weaver nav list to use divs with roles.
This is an alternative approach to hoisting out the list.

This has the least breakage when it comes to structural changes and CSS.

This uses roles to ensure accessibility compliance.

Anything depending on this for the nav list should change CSS (or similar) references of `ul` into `[role=list]` and `li` into `[role=listitem]`.
Only change the `li` that are within the nav list.